requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
boto3==1.17.92
botocore==1.20.92
Django==3.2.4
django-bootstrap-v5==1.0.4
django-bootstrap5==2.0.1
django-ckeditor==6.1.0
django-crispy-forms==1.11.2
django-environ==0.4.5
django-filter==2.4.0
django-fontawesome-5==1.0.18
django-js-asset==1.2.2
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.2.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
